main: Add debugging for hung tests
To allow us to debug hard-to-reproduce "hung test" failures, this commit
adds special handling of test timeouts which, for any still-running
child processes:

- Sends a SIGUSR1 which causes them to print stack traces of all
threads.
- Sends a SIGINT and wait for children processes to exit.

The goals are to provide diagnostics on where child test clients are
stuck, and to allow the ducktape processes to exit, which should allow
CI to gather logs.