add ability to change the test_log_path
result.py:
Added optional attribute in TestResult class
that has a string used as test_log_path
reporter.py:
Added `use_custom_test_log_path` attribute in
SummaryReporter constructor. It is as boolean
to get (if set to True) the test log path from
the result object context and more specifically
from the test_log_path attribute

The combination of the above allows us to set a
different path of our choice for the test logs.
Used when clicking on `Detail` button in the HTML report.
Needed for the ducktape test result combiner
where there will be a new buildkite step that
combines all reports into a single one. Thus developers
will be able to click on the `Detail` button in order
to see the a test result log output by clicking on
the central html summary

ref https://github.com/redpanda-data/devprod/issues/472